refactor(gateway): extract escalation flow into EscalationCoordinator (v0.13 tranche 2)
Adapter shrinks 1640 → 1411 lines (-229); the three escalation
methods + their tight cluster of `pickEscalationPool` / atom-
extractor / thread-marker / synth-LLM dependencies move into a
single 308-line module. 349/349 tests still pass — the v0.13
harness's mra-ask, reaction-approval, and channel @-mention
integration tests transitively exercise the new module without
modification, which is exactly the safety net tranche 1 promised.

### New module `packages/cli/src/gateway/slack/escalation.ts` (308 lines)

`EscalationCoordinator` owns:

- **`escalate()`** (fka `handleEscalation`) — emit the outbound
  `escalate` directive: pick effective pool (filter out asker per
  #30), @-mention or post config-hint when empty, persist
  `pendingSince` marker, append `escalate.triggered` event.
- **`maybeAbsorbReply()`** (fka `maybeAbsorbEscalationReply`) —
  inbound: detect IT contact reply in pending-escalation thread,
  claim the marker, extract atom via LLM, save to pending, post
  the ✅/❌ approval-anchored notice, emit `escalate.absorbed`,
  fire the asker-synthesis follow-up.
- **`postSynthesisedAnswerForAsker()`** (kept as private helper) —
  one-shot LLM call to tag the original asker with the fresh
  knowledge atom answer.

Constructor takes `{ web, config, onLog, llm }`. The
`KnowledgeAtomLike` type moves alongside its only consumer.

### SlackAdapter changes

- Drops 3 methods (~225 lines).
- Drops imports now exclusive to escalation: `pickEscalationPool`,
  `pickEffectiveEscalationPool`, `saveThreadEscalation`,
  `loadThreadEscalation`, `clearThreadEscalation`,
  `extractKnowledgeAtom`, `saveAtom`, `KnowledgeAtomLike`.
- Gains `private readonly escalation: EscalationCoordinator`,
  constructed in the existing DI-aware ctor section.
- Two call sites updated: `this.maybeAbsorbEscalationReply(...)` →
  `this.escalation.maybeAbsorbReply(...)`, `this.handleEscalation(...)` →
  `this.escalation.escalate(...)`.

### Behaviour is byte-equivalent

The 27 v0.13 integration tests cover every branch the extraction
touches: mra-ask escalate path → handleEscalation no-usable-contacts
branch (verified during PR #53 live-Slack); reaction-approval atom
flow → maybeAbsorbReply downstream; channel @-mention → routes
through `this.escalation.maybeAbsorbReply` (no change in
observable outcome). Live-Slack smoke from PR #52 / #53 on
slack-webhook would re-confirm if needed, but tests-only is the
TDD intent here.
